Parking capacity and create fixed

+++ Garage2.0MVC/Models/VehicleType.cs
@@ -9,6 +9,6 @@ namespace Garage2._0MVC.Models
     {
         public int Id { get; set; }
         public VehicleTypeEnum Type { get ; set; }
-        public double Capacity { get; set; }
+        public int Capacity { get; set; }
     }
 }
